moved decls to setup + int -> int8

+++ examples/tests/test1.m
@@ -1,7 +1,7 @@
 define TEST1 2 * 5 + 3
 setup()
 {
-  int test = TEST1;
+  int16 test = TEST1;
 }
 
 mainloop(){}
